Connection opérationnelle et déplacement ok, par contre fermeture serveur réception

+++ code_fork.c
@@ -31,7 +31,6 @@ if (sigaction(SIGUSR1, &action, NULL)) {
 	return -1;
 }
 
-sd = create_socket_by_numbers (NULL, 0, ANONY_SOCK);
 fcntl (tube[0], F_SETFL, O_NONBLOCK);
 
 /* Démarrage de la boucle */
@@ -39,6 +38,8 @@ while (!arret) {
 	int tps_connex = 0;
 	int temp = 0, terminaison_boucle = 0;
 	char c = 0;
+	char buffer[3];
+	sd = create_socket_by_numbers (NULL, 0, ANONY_SOCK);
 	/* Attend SIGUSR1 ou 2 lui indiquant que l'utilisateur a réalisé la demande
 	 * de prise en main */
 	fprintf (stderr, "Test\n");
@@ -52,39 +53,51 @@ while (!arret) {
 			connecte = connect(sd, (struct sockaddr *)&rc_server, sizeof(rc_server));
 		if (!connecte) {
 			fprintf(stderr, "Connexion établie!\n");
-			sleep(1);
 			/* Attends la prise en charge complète du serveur */
-			recv (sd, &tps_connex, sizeof(int), 0);
-			/* Préviens le programme principal de la prise en charge */
-			kill(ppid, SIGUSR1);
+			if (recv (sd, buffer, 3*sizeof(char), 0) > 0) {
+				tps_connex = atoi(buffer);
+				sleep(2);
+				/* Préviens le programme principal de la prise en charge */
+				kill(ppid, SIGUSR1);
 
-			fprintf(stderr, "Pris en charge...\n");
+				fprintf(stderr, "Pris en charge...\n");
 
-			/* Utilisation du contrôle */
-			fcntl (sd, F_SETFL, O_NONBLOCK | O_NDELAY);
-			while (recv(sd, &temp, sizeof(int), 0) && !terminaison_boucle) {
-				usleep(10000);
-				read(tube[0], &c, sizeof(char));
-				switch (c) {
-					case 'u':
-					case 'd':
-					case 'r':
-					case 'l':
-						fprintf(stderr, "Envoi en cours.\n");
-						send (sd, (void *)&c, sizeof(char), 0);
-						break;
-					case 'q':
-						close (sd);
-						terminaison_boucle ++;
-						sommeil = 1;
-						kill(ppid, SIGUSR2);
-						break;
-					default:
-						break;
+				/* Utilisation du contrôle */
+				fcntl (sd, F_SETFL, O_NONBLOCK);
+				while (recv(sd, &temp, sizeof(int), 0) && !terminaison_boucle) {
+					usleep(10000);
+					c = ' ';
+					if (read(tube[0], &c, sizeof(char))!=-1) {
+					switch (c) {
+						case 'u':
+							fprintf (stderr, "Lu u\n");
+							send (sd, (void *)&c, sizeof(char), 0);
+							break;
+						case 'd':
+							fprintf (stderr, "Lu d\n");
+							send (sd, (void *)&c, sizeof(char), 0);
+							break;
+						case 'r':
+							fprintf (stderr, "Lu r\n");
+							send (sd, (void *)&c, sizeof(char), 0);
+							break;
+						case 'l':
+							fprintf(stderr, "Envoi en cours.\n");
+							send (sd, (void *)&c, sizeof(char), 0);
+							break;
+						case 'q':
+							fprintf(stderr, "Réception q\n");
+							terminaison_boucle ++;
+							break;
+						default:
+							break;
+					}
+				}
 				}
 			}
-			sommeil = 1;
 		}
 	}
+	kill(ppid, SIGUSR2);
+	close(sd);
 }
